+++ packages/okta-react/src/withOktaAuth.js
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2017-Present, Okta, Inc. and/or its affiliates. All rights reserved.
+ * Copyright (c) 2020-Present, Okta, Inc. and/or its affiliates. All rights reserved.
  * The Okta software accompanied by this notice is provided pursuant to the Apache License, Version 2.0 (the "License.")
  *
  * You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0.
@@ -10,17 +10,16 @@
  * See the License for the specific language governing permissions and limitations under the License.
  */
 
-import { Component } from 'react';
-import PropTypes from 'prop-types';
+import React from 'react';
+import { useOktaAuth } from './OktaContext';
 
-export default class Secure extends Component {
-  static contextTypes = {
-    auth: PropTypes.object.isRequired
-  }
+const withOktaAuth = (ComponentToWrap) => { 
+  const WrappedComponent = (props) => { 
+    const oktaAuthProps = useOktaAuth();
+    return <ComponentToWrap {...oktaAuthProps } {...props} />;
+  };
+  WrappedComponent.displayName = 'withOktaAuth_' + (ComponentToWrap.displayName || ComponentToWrap.name);
+  return WrappedComponent;
+};
 
-  render() {
-    return this.props.render({
-      auth: this.context.auth
-    });
-  }
-}
+export default withOktaAuth;