+++ packages/okta-react/test/e2e/harness/config-overrides.js
@@ -2,7 +2,9 @@
 
 const webpack = require('webpack');
 const path = require('path');
-const MAIN_ENTRY = path.resolve(__dirname, '../../../dist/index.js');
+const MODULE_DIR = path.resolve(__dirname, '../../../');
+const MAIN_ENTRY = path.join(MODULE_DIR, 'dist/');
+const NODE_MODULES = path.join(MODULE_DIR, 'node_modules');
 
 require('../../../../../.oidc.config.js'); // will load environment vars from testenv
 
@@ -36,7 +38,14 @@ module.exports = {
       config.resolve.alias = {};
     }
     Object.assign(config.resolve.alias, {
-      '@okta/okta-react': MAIN_ENTRY
+      '@okta/okta-react': MAIN_ENTRY,
+
+      // We receive strange errors if there are more than one copy of react modules.
+      // This happens because okta-react has devDependencies installed. Use copy from okta-react.
+      'react': path.join(NODE_MODULES, 'react'),
+      'react-dom': path.join(NODE_MODULES, 'react-dom'),
+      'react-router': path.join(NODE_MODULES, 'react-router'),
+      'react-router-dom': path.join(NODE_MODULES, 'react-router-dom')
     });
 
     config.devtool = 'source-map';