+++ packages/okta-react/test/e2e/harness/src/App.js
@@ -11,45 +11,53 @@
  */
 
 import React, { Component } from 'react';
-import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
-import { Security, SecureRoute, ImplicitCallback, Auth } from '@okta/okta-react';
+import { Route, Switch } from 'react-router-dom';
+import { withRouter } from 'react-router';
+import { AuthService, Security, LoginCallback, SecureRoute } from '@okta/okta-react';
 import Home from './Home';
 import Protected from './Protected';
 import CustomLogin from './CustomLogin';
 import SessionTokenLogin from './SessionTokenLogin';
 
-if (!Auth) {
-  throw new Error('Auth should be defined');
+if (!AuthService) {
+  throw new Error('AuthService should be defined');
 }
 
 class App extends Component {
+  constructor(props) {
+    super(props);
+    this.onAuthRequired = this.onAuthRequired.bind(this);
+  }
+
+  onAuthRequired() {
+    this.props.history.push('/login')
+  }
+  
   render() {
      /* global process */
     const { ISSUER, CLIENT_ID } = process.env;
     const { pkce, redirectUri } = this.props;
     return (
       <React.StrictMode>
-        <Router>
-          <Security issuer={ISSUER}
-                    clientId={CLIENT_ID}
-                    disableHttpsCheck={true}
-                    redirectUri={redirectUri}
-                    onAuthRequired={({history}) => history.push('/login')}
-                    pkce={pkce}>
-            <Switch>
-              <Route path='/login' component={CustomLogin}/>
-              <Route path='/sessionToken-login' component={SessionTokenLogin}/>
-              <SecureRoute exact path='/protected' component={Protected}/>
-              <Route path='/implicit/callback' component={ImplicitCallback} />
-              <Route path='/pkce/callback' component={ImplicitCallback} />
-              <Route path='/' component={Home}/>
-            </Switch>
-          </Security>
-        </Router>
+        <Security issuer={ISSUER}
+                  clientId={CLIENT_ID}
+                  disableHttpsCheck={true}
+                  redirectUri={redirectUri}
+                  onAuthRequired={this.onAuthRequired}
+                  pkce={pkce}>
+          <Switch>
+            <Route path='/login' component={CustomLogin}/>
+            <Route path='/sessionToken-login' component={SessionTokenLogin}/>
+            <SecureRoute exact path='/protected' component={Protected}/>
+            <Route path='/implicit/callback' component={LoginCallback} />
+            <Route path='/pkce/callback' component={LoginCallback} />
+            <Route path='/' component={Home}/>
+          </Switch>
+        </Security>
         <a href="/?pkce=1">PKCE Flow</a> | <a href="/">Implicit Flow</a>
       </React.StrictMode>
     );
   }
 }
 
-export default App;
+export default withRouter(App);