+++ packages/okta-react/test/e2e/harness/src/Protected.js
@@ -11,9 +11,9 @@
  */
 
 import React, { Component } from 'react';
-import { withAuth } from '@okta/okta-react';
+import { withOktaAuth } from '@okta/okta-react';
 
-export default withAuth(class Protected extends Component {
+export default withOktaAuth(class Protected extends Component {
   constructor(props) {
     super(props);
     this.state = { userinfo: null };
@@ -21,13 +21,13 @@ export default withAuth(class Protected extends Component {
   }
 
   async componentDidMount() {
-    const claims = await this.props.auth.getUser();
+    const claims = await this.props.authService.getUser();
     const userinfo = JSON.stringify(claims, null, 4);
     this.setState({ userinfo });
   }
 
   async logout() {
-    this.props.auth.logout('/');
+    this.props.authService.logout('/');
   }
 
   render() {