+++ packages/okta-react/test/jest/secureRoute.test.js
@@ -5,12 +5,87 @@ import SecureRoute from '../../src/SecureRoute';
 import Security from '../../src/Security';
 
 describe('<SecureRoute />', () => {
-  const mockProps = {
-    auth: {
-      isAuthenticated: jest.fn()
-    },
-  };
-
+  let authService;
+  let authState;
+  let mockProps;
+  beforeEach(() => {
+    authState = {
+      isPending: true
+    };
+    authService = {
+      on: jest.fn(),
+      updateAuthState: jest.fn(),
+      getAuthState: jest.fn().mockImplementation(() => authState),
+      login: jest.fn()
+    };
+    mockProps = { authService };
+  });
+  describe('isAuthenticated: true', () => {
+    beforeEach(() => {
+      authState.isAuthenticated = true;
+    });
+    it('will render wrapped component', () => {
+      const MyComponent = function() { return <div>hello world</div>; };
+      const wrapper = mount(
+        <MemoryRouter>
+          <Security {...mockProps}>
+            <SecureRoute
+              component={MyComponent}
+            />
+          </Security>
+        </MemoryRouter>
+      );
+      expect(wrapper.find(MyComponent).html()).toBe('<div>hello world</div>');
+    });
+  });
+  describe('isAuthenticated: false', () => {
+    beforeEach(() => {
+      authState.isAuthenticated = false;
+    });
+    it('will not render wrapped component', () => {
+      const MyComponent = function() { return <div>hello world</div>; };
+      const wrapper = mount(
+        <MemoryRouter>
+          <Security {...mockProps}>
+            <SecureRoute
+              component={MyComponent}
+            />
+          </Security>
+        </MemoryRouter>
+      );
+      expect(wrapper.find(MyComponent).length).toBe(0);
+    });
+    describe('isPending: false', () => {
+      beforeEach(() => {
+        authState.isPending = false;
+      });
+      it('calls login()', () => {
+        mount(
+          <MemoryRouter>
+            <Security {...mockProps}>
+              <SecureRoute />
+            </Security>
+          </MemoryRouter>
+        );
+        expect(authService.login).toHaveBeenCalled();
+      });
+    });
+    describe('isPending: true', () => {
+      beforeEach(() => {
+        authState.isPending = true;
+      });
+      it('does not call login()', () => {
+        mount(
+          <MemoryRouter>
+            <Security {...mockProps}>
+              <SecureRoute />
+            </Security>
+          </MemoryRouter>
+        );
+        expect(authService.login).not.toHaveBeenCalled();
+      });
+    });
+  });
   it('should accept a "path" prop and render a component', () => {
     const wrapper = mount(
       <MemoryRouter>
@@ -23,6 +98,7 @@ describe('<SecureRoute />', () => {
     );
     expect(wrapper.find(SecureRoute).props().path).toBe('/');
   });
+
   it('should accept an "exact" prop and render a component', () => {
     const wrapper = mount(
       <MemoryRouter>
@@ -36,6 +112,7 @@ describe('<SecureRoute />', () => {
     );
     expect(wrapper.find(SecureRoute).props().exact).toBe(true);
   });
+
   it('should not contain an "exact" prop and render a component', () => {
     const wrapper = mount(
       <MemoryRouter>
@@ -48,6 +125,7 @@ describe('<SecureRoute />', () => {
     );
     expect(wrapper.find(SecureRoute).props().exact).toBeUndefined();
   });
+
   it('should accept a "strict" prop and render a component', () => {
     const wrapper = mount(
       <MemoryRouter>
@@ -61,6 +139,7 @@ describe('<SecureRoute />', () => {
     );
     expect(wrapper.find(SecureRoute).props().strict).toBe(true);
   });
+
   it('should accept a "sensitive" prop and render a component', () => {
     const wrapper = mount(
       <MemoryRouter>
@@ -74,6 +153,7 @@ describe('<SecureRoute />', () => {
     );
     expect(wrapper.find(SecureRoute).props().sensitive).toBe(true);
   });
+
   it('should accept an "exact" prop and pass it to an internal Route', () => {
     const wrapper = mount(
       <MemoryRouter>
@@ -88,6 +168,7 @@ describe('<SecureRoute />', () => {
     const secureRoute = wrapper.find(SecureRoute);
     expect(secureRoute.find(Route).props().exact).toBe(true);
   });
+
   it('should accept a "strict" prop and pass it to an internal Route', () => {
     const wrapper = mount(
       <MemoryRouter>
@@ -102,6 +183,7 @@ describe('<SecureRoute />', () => {
     const secureRoute = wrapper.find(SecureRoute);
     expect(secureRoute.find(Route).props().strict).toBe(true);
   });
+
   it('should accept a "sensitive" prop and pass it to an internal Route', () => {
     const wrapper = mount(
       <MemoryRouter>